+++ crates/assembly/src/instruction/ext2_ops.rs
@@ -1,7 +1,7 @@
 use miden_core::{events::SystemEvent::Ext2Inv, operations::Operation::*};
 
 use super::BasicBlockBuilder;
-use crate::{Report, ZERO};
+use crate::ZERO;
 
 /// Given a stack in the following initial configuration [b0, b1, a0, a1, ...] where a = (a0, a1)
 /// and b = (b0, b1) represent elements in the extension field of degree 2, this series
@@ -115,7 +115,7 @@ pub fn ext2_neg(block_builder: &mut BasicBlockBuilder) {
 /// assert b  = (1, 0) | (1, 0) is the multiplicative identity of extension field.
 ///
 /// This operation takes 8 VM cycles.
-pub fn ext2_inv(block_builder: &mut BasicBlockBuilder) -> Result<(), Report> {
+pub fn ext2_inv(block_builder: &mut BasicBlockBuilder) {
     block_builder.push_system_event(Ext2Inv);
     #[rustfmt::skip]
     let ops = [
@@ -129,6 +129,4 @@ pub fn ext2_inv(block_builder: &mut BasicBlockBuilder) -> Result<(), Report> {
         Assert(ZERO),   // [a0', a1', ...] (verify c0 is truthy, i.e. == 1)
     ];
     block_builder.push_ops(ops);
-
-    Ok(())
 }