+++ crates/assembly/src/instruction/mod.rs
@@ -48,7 +48,7 @@ impl Assembler {
         // Start tracking the instruction about to be executed; this will allow us to map the
         // instruction to the sequence of operations which were executed as a part of this
         // instruction.
-        block_builder.track_instruction(instruction, proc_ctx)?;
+        block_builder.track_instruction(instruction, proc_ctx);
 
         // For node-creating instructions, finalize the AssemblyOp now (it will have 0 cycles
         // since no operations have been added yet for this instruction).
@@ -150,9 +150,13 @@ impl Assembler {
             Instruction::Neq => block_builder.push_ops([Eq, Not]),
             Instruction::NeqImm(imm) => field_ops::neq_imm(block_builder, imm.expect_value()),
             Instruction::Lt => field_ops::lt(block_builder),
+            Instruction::LtImm(imm) => field_ops::lt_imm(block_builder, imm.expect_value()),
             Instruction::Lte => field_ops::lte(block_builder),
+            Instruction::LteImm(imm) => field_ops::lte_imm(block_builder, imm.expect_value()),
             Instruction::Gt => field_ops::gt(block_builder),
+            Instruction::GtImm(imm) => field_ops::gt_imm(block_builder, imm.expect_value()),
             Instruction::Gte => field_ops::gte(block_builder),
+            Instruction::GteImm(imm) => field_ops::gte_imm(block_builder, imm.expect_value()),
             Instruction::IsOdd => field_ops::is_odd(block_builder),
 
             // ----- ext2 instructions ------------------------------------------------------------
@@ -161,7 +165,7 @@ impl Assembler {
             Instruction::Ext2Mul => ext2_ops::ext2_mul(block_builder),
             Instruction::Ext2Div => ext2_ops::ext2_div(block_builder),
             Instruction::Ext2Neg => ext2_ops::ext2_neg(block_builder),
-            Instruction::Ext2Inv => ext2_ops::ext2_inv(block_builder)?,
+            Instruction::Ext2Inv => ext2_ops::ext2_inv(block_builder),
 
             // ----- u32 manipulation -------------------------------------------------------------
             Instruction::U32Test => block_builder.push_ops([Dup0, U32split, Drop, Eqz]),
@@ -597,7 +601,7 @@ impl Assembler {
             // ----- debug decorators -------------------------------------------------------------
             Instruction::Debug(options) => {
                 block_builder
-                    .push_decorator(Decorator::Debug(debug::compile_options(options, proc_ctx)?))?;
+                    .push_decorator(Decorator::Debug(debug::compile_options(options, proc_ctx)))?;
             },
 
             Instruction::DebugVar(debug_var_info) => {